+++ awsaccountmgr/organization.py
@@ -2,12 +2,14 @@
 import boto3
 from time import sleep
 from .sts import create_boto3_client
+from .account import Account, AlternateContact
 
 
 class Organization:
+    """Interact with AWS Organization API."""
 
     def __init__(self, root_ou_id, credentials):
-        """Initialise the boto3 organisation session
+        """Initialise the boto3 organisation session.
 
         :param org_client: A 'organizations' boto3 client in the Master account
         :param root_ou_id: The ID of the root Organizational Unit
@@ -17,17 +19,21 @@ class Organization:
         self.root_ou_id = root_ou_id
 
         self._org_client = create_boto3_client(
-            self.get_master_account_id(),
-            'organizations',
-            self._master_credentials
+            self.get_master_account_id(), "organizations", self._master_credentials
+        )
+
+        self._account_client = create_boto3_client(
+            self.get_master_account_id(), "account", self._master_credentials
         )
 
     def list_organizational_units_for_parent(self, parent_ou):
-        """Returns a list of OUs for the given parent"""
+        """Returns a list of OUs for the given parent."""
         organizational_units = [
             ou
-            for org_units in self._org_client.get_paginator("list_organizational_units_for_parent").paginate(ParentId=parent_ou)
-            for ou in org_units['OrganizationalUnits']
+            for org_units in self._org_client.get_paginator(
+                "list_organizational_units_for_parent"
+            ).paginate(ParentId=parent_ou)
+            for ou in org_units["OrganizationalUnits"]
         ]
         return organizational_units
 
@@ -36,31 +42,31 @@ class Organization:
         existing_accounts = [
             account
             for accounts in self._org_client.get_paginator("list_accounts").paginate()
-            for account in accounts['Accounts']
+            for account in accounts["Accounts"]
         ]
         return existing_accounts
 
     def get_account_id(self, account_name):
-        """Retrieves the account ID
+        """Retrieves the account ID.
 
         :param account_name: The name of the account
         :return: Id of account, None if account does not exist
         """
         for account in self.list_accounts():
             if account["Name"].strip() == account_name.strip():
-                return account['Id']
+                return account["Id"]
 
         return None
 
     def get_ou_id(self, ou_path, parent_ou_id=None):
-        """Retrieve an organisational unit id
+        """Retrieve an organisational unit id.
 
         :param ou_path: the name of the organisational unit. For nested ou's use / notation, eg. eu/eudev
         :param parent_ou_id: The parent from where to start parsing the ou_path, defaults to root
         :return: (ou_id, parent_id)
         """
         # Return root OU if '/' is provided
-        if ou_path.strip() == '/':
+        if ou_path.strip() == "/":
             return self.root_ou_id
 
         # Set initial OU to start looking for given ou_path
@@ -68,19 +74,21 @@ class Organization:
             parent_ou_id = self.root_ou_id
 
         # Parse ou_path and find the ID
-        ou_hierarchy = ou_path.strip('/').split('/')
+        ou_hierarchy = ou_path.strip("/").split("/")
         hierarchy_index = 0
 
         while hierarchy_index < len(ou_hierarchy):
             org_units = self.list_organizational_units_for_parent(parent_ou_id)
-            
+
             for ou in org_units:
-                if ou['Name'] == ou_hierarchy[hierarchy_index]:
-                    parent_ou_id = ou['Id']
+                if ou["Name"] == ou_hierarchy[hierarchy_index]:
+                    parent_ou_id = ou["Id"]
                     hierarchy_index += 1
                     break
             else:
-                raise ValueError(f'Could not find ou with name {ou_hierarchy} in OU list {org_units}.')
+                raise ValueError(
+                    f"Could not find ou with name {ou_hierarchy} in OU list {org_units}."
+                )
 
         return parent_ou_id
 
@@ -103,22 +111,26 @@ class Organization:
 
         # TODO error handling and check if an account always returns just one item
         response = self._org_client.list_parents(ChildId=account_id)
-        source_parent_id = response['Parents'][0]['Id']
+        source_parent_id = response["Parents"][0]["Id"]
 
         if source_parent_id == ou_id:
             # Account is already resided in ou_path
             return
 
-        if allow_direct_move is False and source_parent_id != self.root_ou_id and ou_id != self.root_ou_id:
+        if (
+            allow_direct_move is False
+            and source_parent_id != self.root_ou_id
+            and ou_id != self.root_ou_id
+        ):
             raise ValueError(
                 f"Trying to move an account from non-root OU to another OU. "
                 "Move account to root ({self.root_ou_id}) first."
             )
 
-        response = self._org_client.move_account(
+        self._org_client.move_account(
             AccountId=account_id,
             SourceParentId=source_parent_id,
-            DestinationParentId=ou_id
+            DestinationParentId=ou_id,
         )
 
     def create_account_tags(self, account_id, tags):
@@ -131,37 +143,60 @@ class Organization:
         # meaning that any tag not referenced will be removed
         # from the account. use org_client.untag_resource()
         formatted_tags = [
-            {'Key': key, 'Value': value}
-            for tag in tags
-            for key, value in tag.items()
+            {"Key": key, "Value": value} for tag in tags for key, value in tag.items()
         ]
-        self._org_client.tag_resource(
-            ResourceId=account_id,
-            Tags=formatted_tags
-        )
+        self._org_client.tag_resource(ResourceId=account_id, Tags=formatted_tags)
 
     @staticmethod
     def get_master_account_id():
-        """Retrieves the master account id from API"""
-        org_client = boto3.client('organizations')
+        """Retrieves the master account id from API."""
+        org_client = boto3.client("organizations")
         response = org_client.describe_organization()
-        return response['Organization']['MasterAccountId']
+        return response["Organization"]["MasterAccountId"]
 
     def create_account_alias(self, account_id, account_alias):
-        """Creates or updates the account alias for given account_id"""
+        """Creates or updates the account alias for given account_id."""
 
-        iam_client = create_boto3_client(
-            account_id,
-            "iam",
-            self._master_credentials
-        )
+        iam_client = create_boto3_client(account_id, "iam", self._master_credentials)
 
         try:
             iam_client.create_account_alias(AccountAlias=account_alias)
         except iam_client.exceptions.EntityAlreadyExistsException:
             pass  # Alias already exists
 
-    def create_account(self, account):
+    def remove_alternate_contact(self, account_id, contact_type):
+        """Removes the contact type from the account."""
+
+        if contact_type.upper() not in ["BILLING", "OPERATIONS", "SECURITY"]:
+            raise ValueError(f"Contact type {contact_type} is not supported.")
+
+        self._account_client.delete_alternate_contact(
+            AccountId=account_id, AlternateContactType=contact_type.upper()
+        )
+
+    def update_alternate_contact(
+        self, account_id: str, contact_type: str, contact_details: AlternateContact
+    ):
+        """Create or update the account alternate contacts for given account.
+
+        :param account_id: ID of the AWS account
+        :param contact_type: Type of contact to update
+        :param contact_details: AlternateContact class instance
+        """
+
+        if contact_type.upper() not in ["BILLING", "OPERATIONS", "SECURITY"]:
+            raise ValueError(f"Contact type {contact_type} is not supported.")
+
+        self._account_client.put_alternate_contact(
+            AccountId=account_id,
+            AlternateContactType=contact_type,
+            EmailAddress=contact_details.email,
+            Name=contact_details.name,
+            PhoneNumber=contact_details.phone,
+            Title=contact_details.title,
+        )
+
+    def create_account(self, account: Account):
         """Creates a new the account if it doesn't exist.
 
         :param account: Class instance of Account
@@ -180,7 +215,7 @@ class Organization:
                     'AccountId': '173289020951',
                     'FailureReason': 'ACCOUNT_LIMIT_EXCEEDED'
                 }
-            """
+        """
         if account.allow_billing is True:
             allow_billing = "ALLOW"
         else:
@@ -210,7 +245,3 @@ class Organization:
         sleep(10)  # Wait until OrganizationalRole is created in new account
 
         return account_id
-
-
-if __name__ == "__main__":
-    pass